+++ src/main/java/com/atlauncher/gui/LauncherFrame.java
@@ -21,6 +21,7 @@ import com.atlauncher.App;
 import com.atlauncher.LogManager;
 import com.atlauncher.data.Constants;
 import com.atlauncher.data.Pack;
+import com.atlauncher.data.PackVersion;
 import com.atlauncher.evnt.listener.RelocalizationListener;
 import com.atlauncher.evnt.manager.RelocalizationManager;
 import com.atlauncher.evnt.manager.TabChangeManager;
@@ -106,6 +107,28 @@ public final class LauncherFrame extends JFrame implements RelocalizationListene
                 LogManager.error("Error automatically installing " + (pack == null ? "pack" : pack.getName()) + "!");
             } else {
                 new InstanceInstallerDialog(pack);
+            }
+        } else if (App.packShareCodeToInstall != null) {
+            String[] parts = App.packShareCodeToInstall.split("|||");
+
+            if (parts.length != 3) {
+                LogManager.error("Error automatically installing pack from share code!");
+            } else {
+                Pack pack = App.settings.getPackBySafeName(parts[0]);
+
+                if (pack == null) {
+                    LogManager.error("Error automatically installing pack from share code!");
+                } else {
+                    PackVersion version = pack.getVersionByName(parts[1]);
+
+                    if (version == null) {
+                        LogManager.error("Error automatically installing " + pack.getName() + " from share code!");
+                    }
+                    String shareCode = parts[2];
+
+                    new InstanceInstallerDialog(pack, version, shareCode);
+                }
+
             }
         }
     }